added texture resavability and fixed goodcollide
added a feature where if the texture is animated and it switched to a texture that it had been before it can load that texture rather than reloading the other one from path, which is much faster. i also fixed a few small errors in the goodcollide() function that was causing it to not work.